обавлена логика обработки ошибок экрана "Вакансия"

+++ app/src/main/java/ru/practicum/android/diploma/vacancy/data/VacancyRepositoryImpl.kt
@@ -5,8 +5,8 @@ import kotlinx.coroutines.flow.flow
 import ru.practicum.android.diploma.favourites.domain.api.FavouritesRepository
 import ru.practicum.android.diploma.search.data.dto.RESULT_CODE_NOT_FOUND
 import ru.practicum.android.diploma.search.data.dto.RESULT_CODE_NO_INTERNET
-import ru.practicum.android.diploma.search.data.dto.RESULT_CODE_SUCCESS
 import ru.practicum.android.diploma.search.data.network.NetworkClient
+import ru.practicum.android.diploma.search.domain.utils.ResponseData
 import ru.practicum.android.diploma.vacancy.data.dto.VacancyRequest
 import ru.practicum.android.diploma.vacancy.data.dto.VacancyResponse
 import ru.practicum.android.diploma.vacancy.data.dto.responseToVacancyFull
@@ -18,25 +18,32 @@ class VacancyRepositoryImpl(
     private val favouritesRepository: FavouritesRepository,
 ) : VacancyRepository {
 
-    override fun getVacancy(id: Int): Flow<VacancyFull> = flow {
-        val response = networkClient.doRequest(VacancyRequest(id = id))
-        when (response.resultCode) {
-            RESULT_CODE_SUCCESS -> {
-                val vacancyFull = responseToVacancyFull(response as VacancyResponse)
+    override fun getVacancy(id: Int): Flow<ResponseData<VacancyFull>> = flow {
+        when (val response = networkClient.doRequest(VacancyRequest(id))) {
+            is VacancyResponse -> {
+                val vacancyFull = responseToVacancyFull(response)
                 favouritesRepository.updateVacancy(vacancyFull)
-                emit(vacancyFull)
+                emit(ResponseData.Data(vacancyFull))
             }
 
-            RESULT_CODE_NO_INTERNET -> {
-                emit(favouritesRepository.getById(id) ?: VacancyFull())
-            }
+            else -> {
+                when (response.resultCode) {
+                    RESULT_CODE_NO_INTERNET -> {
+                        if (favouritesRepository.getById(id) != null) {
+                            emit(ResponseData.Data(favouritesRepository.getById(id)!!))
+                        } else {
+                            emit(ResponseData.Error(ResponseData.ResponseError.NO_INTERNET))
+                        }
+                    }
 
-            RESULT_CODE_NOT_FOUND -> {
-                favouritesRepository.deleteVacancy(id)
-                emit(VacancyFull())
-            }
+                    RESULT_CODE_NOT_FOUND -> {
+                        favouritesRepository.deleteVacancy(id)
+                        emit(ResponseData.Error(ResponseData.ResponseError.NOT_FOUND))
+                    }
 
-            else -> emit(VacancyFull())
+                    else -> emit(ResponseData.Error(ResponseData.ResponseError.SERVER_ERROR))
+                }
+            }
         }
     }
 }
